% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{choose_center}
\alias{choose_center}
\title{Choose the center with the largest gradient norm}
\usage{
choose_center(x, x_star, params)
}
\arguments{
\item{x}{coordinates of the position (numeric vector of length 2)}

\item{x_star}{matrix of coordinates of each center in the mixture (J x 2)}

\item{params}{List of parameters:
- alpha: list of weights in the mixture
- B:  list of precision matrices in the mixture}
}
\value{
index of the chosen center
}
\description{
Choose the center with the largest gradient norm
}
