% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{closest_point_on_boundary}
\alias{closest_point_on_boundary}
\title{Find the closest point on the boundary of a single polygon}
\usage{
closest_point_on_boundary(x, polygon, grad = TRUE)
}
\arguments{
\item{x}{Numeric vector of length 2: coordinates of the point}

\item{polygon}{Polygon object (with @coords slot)}

\item{grad}{logical. If TRUE, also compute the gradient of the projection}
}
\value{
A list with elements:
 - point: Numeric vector of length 2, coordinates of the closest boundary point
 - gradient: 2x2 Jacobian of the projection (or NULL at vertices)
}
\description{
Find the closest point on the boundary of a single polygon
}
