% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{compute_langevin_weight}
\alias{compute_langevin_weight}
\title{Helper function to compute particle weights}
\usage{
compute_langevin_weight(
  U_pred,
  U_prev,
  y,
  M,
  delta,
  push,
  potential_params = NULL,
  tau,
  nu,
  omega,
  error_dist,
  error_params,
  scheme = "Lie-Trotter",
  ind_fixed_point = NULL,
  L = NULL,
  Q = NULL,
  proposal_weight = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{U_pred}{Numeric vector of length 4. Predicted state vector
\eqn{(x, y, v_x, v_y)} representing position and velocity.}

\item{U_prev}{Numeric vector of length 4. Previous state vector}

\item{y}{Numeric vector of length 2. Observed position at the next time step.}

\item{M}{Numeric matrix. Observation link matrix.}

\item{delta}{Numeric scalar. Time step size.}

\item{push}{Numeric vector. External push (control or forcing term).}

\item{potential_params}{Parameters for mixture of gaussian potential. List
with elements x_star, B and alpha}

\item{tau}{Numeric scalar. Correlation timescale parameter.}

\item{nu}{Numeric scalar. Velocity  parameter.}

\item{omega}{Numeric scalar. Rotation parameter.}

\item{error_dist}{Character string. Distribution of observation error. 
One of `"scaled_t"`, `"normal"`, or `"argos"`.}

\item{error_params}{List of error distribution parameters:
\describe{
  \item{`scale`, `df`}{(for `"scaled_t"`) scale and degrees of freedom.}
  \item{`sigma_obs`}{(for `"normal"` and `"argos"`) observation error scale.}
  \item{`df`, `rho`, `a`, `p`}{(for `"argos"`) degrees of freedom, 
    correlation, anisotropy parameter, and mixture probability.}
}}

\item{scheme}{Character string. Numerical scheme for propagation, 
either `"Lie-Trotter"` (default) or `"Strang"`.}

\item{ind_fixed_point}{index of fixed point of the mixture of
Gaussian potentials for splitting scheme. Default NULL}

\item{L}{Precomputed observation link matrix for solving the SDE. Default NULL}

\item{Q}{Precomputed process covariance matrix for solving the SDE. Default NULL}

\item{proposal_weight}{weight in the gaussian proposal mean and covariance}

\item{verbose}{Logical, whether to print progress messages (default: FALSE)}
}
\value{
Numeric vector of length 4. The propagated state \eqn{U_{t+\Delta}}
}
\description{
Helper function to compute particle weights
}
