% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{compute_push}
\alias{compute_push}
\title{Compute penalisation term}
\usage{
compute_push(x, polygon, lambda)
}
\arguments{
\item{x}{Numeric vector of length 2: coordinates of the point}

\item{polygon}{SpatialPolygon object representing the polygon}

\item{lambda}{penalty parameter}
}
\value{
Numeric vector of length 2: penalisation term
}
\description{
Compute penalisation term
}
