% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_push_cpp}
\alias{compute_push_cpp}
\title{Compute penalisation term (C++ implementation)}
\usage{
compute_push_cpp(x, coords, lambda)
}
\arguments{
\item{x}{Numeric vector of length 2: coordinates of the point}

\item{coords}{Matrix of polygon boundary coordinates (as in \code{polygon@coords})}

\item{lambda}{penalty parameter}
}
\value{
Numeric vector of length 2: penalisation term
}
\description{
Compute penalisation term (C++ implementation)
}
