% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_models.R
\name{dmvt_mixture}
\alias{dmvt_mixture}
\title{Density of mixture of multivariate student for error in ARGOS positions}
\usage{
dmvt_mixture(x, mean, params, log = FALSE)
}
\arguments{
\item{x}{point where to evaluate the density}

\item{mean}{mean of the student distributions}

\item{params}{parameters a, df, sigma_obs, rho and p in the distribution (see Brost et al. 2015).}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
numeric density value
}
\description{
Density of mixture of multivariate student for error in ARGOS positions
}
