% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gradient.R
\name{dRACVM_link}
\alias{dRACVM_link}
\title{Function to compute partial derivatives of RACVM link matrix T (Delporte et .al 2025)}
\usage{
dRACVM_link(tau, omega, h)
}
\arguments{
\item{tau}{RACVM parameter tau}

\item{omega}{RACVM parameter omega}

\item{h}{time step}
}
\value{
List with three elements: derivatives of Q with respect to
#' tau, nu, and omega : three 4x4 matrices: dT_dtau, dT_dnu = 0, dQ_domega
}
\description{
Function to compute partial derivatives of RACVM link matrix T (Delporte et .al 2025)
}
