% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_models.R
\name{dscaledt}
\alias{dscaledt}
\title{Density for scaled t-distribution used in Fastloc-GPS errors}
\usage{
dscaledt(y, mean, scale, df, log = FALSE)
}
\arguments{
\item{y}{point where we want to evaluate the density}

\item{mean}{mean parameter}

\item{scale}{positive scale parameter}

\item{df}{degree of freedom}

\item{log}{logical; if TRUE, probabilities p are given as log(p)}
}
\value{
numeric
}
\description{
Density for scaled t-distribution used in Fastloc-GPS errors
}
