% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kalman_filter.R
\name{extended_kalman_filter}
\alias{extended_kalman_filter}
\title{Extended Kalman filter algorithm for penalised Langevin SDE 
based on Lie-Trotter approximation}
\usage{
extended_kalman_filter(
  data,
  sde_params,
  potential_hessian,
  potential_params = NULL,
  sigma_obs,
  polygon,
  lambda = Inf,
  U0,
  R0,
  split_around_fixed_point = FALSE
)
}
\arguments{
\item{data}{data frame with columns Y1,Y2 for observations, time for time}

\item{sde_params}{List of parameters tau,nu,omega}

\item{potential_hessian}{Function to compute Hessian matrix of the potential surface}

\item{potential_params}{Parameters for a mixture of gaussian potential. Only necessary
if split_around_fixed_point is TRUE.}

\item{sigma_obs}{Array of covariance matrices for the 
measurement error for each observation}

\item{polygon}{polygon defining the ecological domain of interest}

\item{lambda}{Penalisation parameter. Default : Inf (no penalisation)}

\item{U0}{Initial guess for the hidden state. A vector of length 4 with position and velocity for RACVM, 
and a vector of length for OU}

\item{R0}{Initial guess for the covariance matrix of the hidden state. Dimensions 
4x4 for RACVM and 2x2 for OU}

\item{split_around_fixed_point}{Boolean. If TRUE, split around fixed point of
the potential. Only works for mixture of mixture of gaussian potential}
}
\value{
matrix with filtered states. Dimension (\code{nrow(data)},4)
}
\description{
Extended Kalman filter algorithm for penalised Langevin SDE 
based on Lie-Trotter approximation
}
