% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_smoother.R
\name{forward_filtering_backward_sampling}
\alias{forward_filtering_backward_sampling}
\title{Forward filtering backward Sampling for penalised Langevin SDE 
see Doucet et .al (2011) for the general algorithm.
Currently, works for Strang but only with naive splitting}
\usage{
forward_filtering_backward_sampling(
  data,
  n_samples,
  forward_filter = NULL,
  sde_params,
  potential_params = NULL,
  error_params,
  error_dist,
  polygon,
  U0,
  lambda,
  num_particles,
  scheme = "Lie-Trotter",
  split_around_fixed_point = FALSE,
  ESS_threshold = 1,
  proposal_weight = 0.5,
  verbose = FALSE,
  print_timing = FALSE,
  obs_error_params = NULL
)
}
\arguments{
\item{data}{A data.frame containing observations with columns: time, Y1, Y2}

\item{n_samples}{number of smoothed trajectories to sample}

\item{forward_filter}{output of the particle_filter2D function. If NULL, the function
will run the particle filter first.}

\item{sde_params}{list with elements tau, nu, omega}

\item{potential_params}{parameters for the potential gradient function}

\item{error_params}{parameters for the measurement error distribution}

\item{error_dist}{type of measurement error : "normal", "scaled_t" or "argos"}

\item{polygon}{matrix with two columns defining the vertices of the polygonal domain}

\item{U0}{initial state (vector of length 4)}

\item{lambda}{penalty parameter for the inward push}

\item{num_particles}{number of particles to use in the particle filter}

\item{scheme}{type of splitting scheme to apply "Lie-Trotter" or "Strang"}

\item{split_around_fixed_point}{logical. If TRUE, split the SDE around the fixed point}

\item{ESS_threshold}{threshold for effective sample size to trigger 
resampling in the particle filter}

\item{proposal_weight}{weight for the proposal distribution 
in the particle filter (between 0 and 1)}

\item{verbose}{logical. If TRUE, print information during the computation}

\item{print_timing}{logical. If TRUE, print timing information for each step}
}
\value{
array of dimension (n_samples, num_states, N) 
with the smoothed trajectories7
}
\description{
Forward filtering backward Sampling for penalised Langevin SDE 
see Doucet et .al (2011) for the general algorithm.
Currently, works for Strang but only with naive splitting
}
