% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{get_PF_mean}
\alias{get_PF_mean}
\title{Get estimate of the trajectory (conditional mean) from particle filter}
\usage{
get_PF_mean(PF_results)
}
\arguments{
\item{PF_results}{List as returned by \code{particle_filter_2D}}
}
\value{
a matrix of estimated filtered states based on conditional mean
}
\description{
Get estimate of the trajectory (conditional mean) from particle filter
}
