% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{is_point_inside_polygon}
\alias{is_point_inside_polygon}
\title{Function to check if a point is inside the polygon}
\usage{
is_point_inside_polygon(x, polygon)
}
\arguments{
\item{x}{Numeric vector of length 2: coordinates of the point}

\item{polygon}{SpatialPolygon object representing the polygon}
}
\value{
logical: TRUE if the point is inside or on the edge, FALSE otherwise
}
\description{
Function to check if a point is inside the polygon
}
