% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LangevinAnimalTracking_package.R
\docType{package}
\name{LangevinAnimalTracking-package}
\alias{LangevinAnimalTracking}
\alias{LangevinAnimalTracking-package}
\title{LangevinAnimalTracking: A package for modeling animal movement using Langevin diffusion models}
\description{
Implements filtering algorithms for animal movement trajectories.
}
\author{
\strong{Maintainer}: Alexandre Delporte \email{alexandre.delporte@univ-grenoble-alpes.fr}

Authors:
\itemize{
  \item Susanne Ditlevsen
  \item Adeline Samson
}

}
