% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk.R
\name{llk}
\alias{llk}
\title{Joint log-likelihood for the full data set}
\usage{
llk(
  data,
  push_mat,
  potential_params = NULL,
  tau,
  nu,
  omega,
  error_dist = NULL,
  error_params = NULL,
  scheme = "Lie-Trotter",
  split_around_fixed_point = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{data frame with columns time, Y1, Y2, X1, X2, V1, V2}

\item{push_mat}{matrix with two columns with the inward push values at each time step}

\item{potential_params}{Parameters for a mixture of Gaussian potential. Only necessary
if split_around_fixed_point is TRUE.}

\item{tau}{persistence parameter tau of the SDE}

\item{nu}{speed parameter nu of the SDE}

\item{omega}{rotation parameter omega of the SDE}

\item{error_dist}{Type of measurement error : "normal", "scaled_t" or "argos"}

\item{error_params}{List of parameters for the measurement error distribution}

\item{scheme}{Type of splitting scheme to apply "Lie-Trotter" or "Strang"}

\item{split_around_fixed_point}{logical. If TRUE, split the SDE around the fixed point}

\item{verbose}{logical. If TRUE, print information during the computation}
}
\value{
vector of log-likelihood values (numeric) for each time step
}
\description{
Joint log-likelihood for the full data set
}
