% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk.R
\name{llk_one_step}
\alias{llk_one_step}
\title{Joint log-likelihood for one step of the SDE}
\usage{
llk_one_step(
  Y,
  U_next,
  U_prev,
  delta,
  push_prev,
  push_next = NULL,
  potential_params = NULL,
  tau,
  nu,
  omega,
  error_dist = NULL,
  error_params = NULL,
  scheme = "Lie-Trotter",
  ind_fixed_point = NULL,
  verbose = FALSE
)
}
\arguments{
\item{Y}{numeric vector of observation (length 2)}

\item{U_next}{numeric vector c(X1,X2,V1,V2) of next hidden state}

\item{U_prev}{numeric vector c(X1,X2,V1,V2) of previous hidden state}

\item{delta}{time step between U_prev and U_next}

\item{push_prev}{vector of length 2 with the value of the inward push at time 
of U_prev}

\item{push_next}{vector of length 2 with the value of the inward push at time
of U_next. Only necessary for Strang scheme}

\item{potential_params}{Parameters for a mixture of Gaussian potential.}

\item{tau}{persistence parameter tau of the SDE}

\item{nu}{speed parameter nu of the SDE}

\item{omega}{rotation parameter omega of the SDE}

\item{error_dist}{Type of measurement error : "normal", "scaled_t" or "argos"}

\item{error_params}{List of parameters for the measurement error distribution}

\item{scheme}{Type of splitting scheme to apply "Lie-Trotter" or "Strang"}

\item{ind_fixed_point}{Default NULL. 
If not NULL, index of the fixed point around which to split}

\item{verbose}{logical. If TRUE, print information during the computation}
}
\value{
log-likelihood value (numeric)
}
\description{
Joint log-likelihood for one step of the SDE
}
