% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{log_dmvnorm_chol}
\alias{log_dmvnorm_chol}
\title{Log-density of multivariate normal using Cholesky decomposition}
\usage{
log_dmvnorm_chol(x, mean, cholSigma)
}
\arguments{
\item{x}{Numeric vector of length d: point where to evaluate the density}

\item{mean}{Numeric vector of length d: mean of the distribution}

\item{cholSigma}{Upper Cholesky factor of the covariance matrix Sigma (d x d)}
}
\value{
Log-density value (numeric)
}
\description{
Log-density of multivariate normal using Cholesky decomposition
}
