% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{log_dmvt_chol}
\alias{log_dmvt_chol}
\title{Log-density of multivariate t-distribution using Cholesky decomposition}
\usage{
log_dmvt_chol(x, mean, chol, df)
}
\arguments{
\item{x}{Numeric vector of length d: point where to evaluate the density}

\item{mean}{Numeric vector of length d: mean of the distribution}

\item{chol}{Upper Cholesky factor of the scale matrix (d x d)}

\item{df}{Degrees of freedom}
}
\value{
Log-density value (numeric)
}
\description{
Log-density of multivariate t-distribution using Cholesky decomposition
}
