% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{max_error}
\alias{max_error}
\title{Function to compute max error between true and filtered positions}
\usage{
max_error(true, filtered)
}
\arguments{
\item{true}{matrix of true positions (n x 2)}

\item{filtered}{matrix of filtered positions (n x 2)}
}
\value{
max error value (numeric)
}
\description{
Function to compute max error between true and filtered positions
}
