% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mix_gaussian_grad}
\alias{mix_gaussian_grad}
\title{Compute the gradient of mixture of Gaussian potentials}
\usage{
mix_gaussian_grad(
  x,
  x_star = matrix(c(0, 0), ncol = 2),
  params,
  exclude = NULL
)
}
\arguments{
\item{x}{coordinates of the position}

\item{x_star}{matrix of coordinates of each center in the mixture (default = 0,0)}

\item{params}{List of parameters:
- alpha: list of weights in the mixture
- B:  list of covariance matrices in the mixture}

\item{exclude}{vector of indices of mixture components to exclude from the computation}
}
\value{
A numeric vector of length 2: (dphi/dx, dphi/dy)
}
\description{
Compute the gradient of mixture of Gaussian potentials
}
