% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mix_gaussian_hessian}
\alias{mix_gaussian_hessian}
\title{Compute the Hessian of a Gaussian mixture potential}
\usage{
mix_gaussian_hessian(x, x_star, params, exclude = NULL)
}
\arguments{
\item{x}{Numeric vector of length 2: coordinates of the position}

\item{x_star}{Matrix of size m x 2 giving the coordinates of each center}

\item{params}{List of parameters:
- alpha: numeric vector of length m, weights of the mixture
- B:     list of length m, each element is a 2x2 positive-definite matrix}

\item{exclude}{vector of indices of mixture components to exclude 
from the computation}
}
\value{
A 2x2 Hessian matrix evaluated at x
}
\description{
Compute the Hessian of a Gaussian mixture potential
}
