% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{mix_gaussian_potential}
\alias{mix_gaussian_potential}
\title{Compute value of a mixture of Gaussian potentials}
\usage{
mix_gaussian_potential(x, x_star, params, exclude = NULL)
}
\arguments{
\item{x}{coordinates of the position}

\item{x_star}{matrix of coordinates of each center in the mixture}

\item{params}{List of parameters:
- alpha: list of weights in the mixture
- B:  list of covariance matrices in the mixture}

\item{exclude}{vector of indices of mixture components to exclude from the computation}
}
\value{
A numeric vector of length 2: (dphi/dx, dphi/dy)
}
\description{
Compute value of a mixture of Gaussian potentials
}
