% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{OU_cov_approx}
\alias{OU_cov_approx}
\title{Computation of the OU covariance matrix with the
approximation based on Taylor expansion (Pilipovic, 2025)}
\usage{
OU_cov_approx(Atilde, Gamma, dt)
}
\arguments{
\item{Atilde}{drift matrix}

\item{Gamma}{diffusion matrix}

\item{dt}{time step}
}
\value{
Covariance matrix Q (4x4 matrix)
}
\description{
Computation of the OU covariance matrix with the
approximation based on Taylor expansion (Pilipovic, 2025)
}
