% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{OU_cov_exact}
\alias{OU_cov_exact}
\title{Computation of OU covariance matrix based on kronecker sum (Albertsen, 2019)}
\usage{
OU_cov_exact(A, Gamma, h, expAh)
}
\arguments{
\item{A}{drift matrix}

\item{Gamma}{diffusion matrix}

\item{h}{time step}

\item{expAh}{precomputed matrix exponential of A*h}
}
\value{
Covariance matrix Q (4x4 matrix)
}
\description{
Computation of OU covariance matrix based on kronecker sum (Albertsen, 2019)
}
