% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OU_cov_exact_cpp}
\alias{OU_cov_exact_cpp}
\title{Compute exact covariance matrix for Ornstein-Uhlenbeck process (C++ version)}
\usage{
OU_cov_exact_cpp(A, Gamma, h, expAh)
}
\arguments{
\item{A}{Drift matrix (d x d)}

\item{Gamma}{Process covariance matrix (d x d)}

\item{h}{Time step}

\item{expAh}{Matrix exponential exp(A*h) (d x d)}
}
\value{
Covariance matrix Q (d x d)
}
\description{
Compute exact covariance matrix for Ornstein-Uhlenbeck process (C++ version)
}
