% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OU_cov_exact_cpp_full}
\alias{OU_cov_exact_cpp_full}
\title{Compute exact covariance matrix with matrix exponential computed internally}
\usage{
OU_cov_exact_cpp_full(A, Gamma, h)
}
\arguments{
\item{A}{Drift matrix (d x d)}

\item{Gamma}{Process covariance matrix (d x d)}

\item{h}{Time step}
}
\value{
Covariance matrix Q (d x d)
}
\description{
Compute exact covariance matrix with matrix exponential computed internally
}
