% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{particle_cloud}
\alias{particle_cloud}
\title{Plot cloud of particles to see where particles are landing 
at a specific time step}
\usage{
particle_cloud(t0, data, particles, weights, polygon = NULL)
}
\arguments{
\item{t0}{time}

\item{data}{data frame of observations with columns Y1 and Y2}

\item{particles}{array of particles as returned by \code{particle_filter2D}}

\item{weights}{matrix of particle weights as returned by \code{particle_filter2D}}

\item{polygon}{SpatialPolygons object defining the boundary (from `sp` package`)}
}
\description{
Plot cloud of particles to see where particles are landing 
at a specific time step
}
