% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_models.R
\name{plot_dmvt_mixture}
\alias{plot_dmvt_mixture}
\title{Plot heatmap of mixture of multivariate Student-t density}
\usage{
plot_dmvt_mixture(mean, params, xlim, ylim, n = 100)
}
\arguments{
\item{mean}{vector of length 2: mean of the Student-t distributions}

\item{params}{list with elements a, df, sigma_obs, rho, p}

\item{xlim, ylim}{numeric vectors of length 2 defining plotting limits}

\item{n}{number of grid points per dimension (default = 100)}
}
\value{
ggplot heatmap of the mixture density
}
\description{
Plot heatmap of mixture of multivariate Student-t density
}
