% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{product_gaussian}
\alias{product_gaussian}
\title{Combine two Gaussian distributions into a single Gaussian}
\usage{
product_gaussian(P1, P2, mean1, mean2, M)
}
\arguments{
\item{P1}{Precision matrix (inverse covariance) of dimension \eqn{n \times n}.}

\item{P2}{Precision matrix (inverse covariance) of dimension \eqn{q \times q}.}

\item{mean1}{Numeric vector of length \eqn{n}. Mean of the first Gaussian.}

\item{mean2}{Numeric vector of length \eqn{q}. Mean of the second Gaussian.}

\item{M}{Link matrix of dimension \eqn{q \times n} that maps the state space.}
}
\value{
A list with components:
\describe{
  \item{\code{mean}}{Posterior mean vector.}
  \item{\code{cov}}{Posterior covariance matrix.}
}
}
\description{
Computes the mean and covariance of the product of two Gaussian densities:  
\deqn{N(mean1, P1^{-1}) \times N(L mean2, P2^{-1}) \propto N(m, \Gamma).}
}
