% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{propagate_langevin_particle}
\alias{propagate_langevin_particle}
\title{Helper function to propagate particles for Langevin SDE}
\usage{
propagate_langevin_particle(
  U,
  y,
  M,
  delta,
  push,
  potential_params,
  tau,
  nu,
  omega,
  lambda,
  error_dist,
  error_params,
  scheme = "Lie-Trotter",
  polygon,
  ind_fixed_point = NULL,
  L = NULL,
  Q = NULL,
  verbose = FALSE
)
}
\arguments{
\item{U}{Numeric vector of length 4. Current state vector, typically 
\eqn{(x, y, v_x, v_y)} representing position and velocity.}

\item{y}{Numeric vector of length 2. Observed position at the next time step.}

\item{M}{Numeric matrix. Observation link matrix.}

\item{delta}{Numeric scalar. Time step size.}

\item{push}{Numeric vector. External push .}

\item{potential_params}{Parameters for mixture of gaussian potential. List
with elements x_star, B and alpha}

\item{tau}{Numeric scalar. Correlation timescale parameter.}

\item{nu}{Numeric scalar. Velocity  parameter.}

\item{omega}{Numeric scalar. Rotation parameter.}

\item{lambda}{penalization parameter}

\item{error_dist}{Character string. Distribution of observation error. 
One of `"scaled_t"`, `"normal"`, or `"argos"`.}

\item{error_params}{List of error distribution parameters:
\describe{
  \item{`scale`, `df`}{(for `"scaled_t"`) scale and degrees of freedom.}
  \item{`sigma_obs`}{(for `"normal"` and `"argos"`) observation error scale.}
  \item{`df`, `rho`, `a`, `p`}{(for `"argos"`) degrees of freedom, 
    correlation, anisotropy parameter, and mixture probability.}
}}

\item{scheme}{Character string. Numerical scheme for propagation, 
either `"Lie-Trotter"` (default) or `"Strang"`.}

\item{polygon}{Polygon defining the domain of interest}

\item{ind_fixed_point}{index of fixed point of the mixture of 
Gaussian potentials for splitting scheme. Default NULL}

\item{L}{Precomputed observation link matrix for solving the SDE. Default NULL}

\item{Q}{Precomputed process covariance matrix for solving the SDE. Default NULL}

\item{verbose}{Logical, whether to print progress messages (default: FALSE)}
}
\value{
Numeric vector of length 4. The propagated state \eqn{U_{t+\Delta}}
}
\description{
Helper function to propagate particles for Langevin SDE
}
