% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{RACVM_cov}
\alias{RACVM_cov}
\title{Compute the covariance matrix for the RACVM model}
\usage{
RACVM_cov(tau, nu, omega, dt)
}
\arguments{
\item{tau}{time scale parameter}

\item{nu}{scale parameter}

\item{omega}{angular velocity parameter}

\item{dt}{time step}
}
\value{
Covariance matrix Q (4x4 matrix)
}
\description{
Compute the covariance matrix for the RACVM model
}
