% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{RACVM_link}
\alias{RACVM_link}
\title{Compute the link matrix for the RACVM model}
\usage{
RACVM_link(tau, omega, dt)
}
\arguments{
\item{tau}{time scale parameter}

\item{omega}{angular velocity parameter}

\item{dt}{time step}
}
\value{
Link matrix L (4x4 matrix)
}
\description{
Compute the link matrix for the RACVM model
}
