% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RACVM_link_cpp}
\alias{RACVM_link_cpp}
\title{Compute RACVM link matrix (C++ version)}
\usage{
RACVM_link_cpp(tau, omega, dt)
}
\arguments{
\item{tau}{Time scale parameter}

\item{omega}{Angular velocity parameter}

\item{dt}{Time step}
}
\value{
Link matrix L (4x4 matrix)
}
\description{
Computes the state transition matrix for the RACVM model.
}
