% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rmse}
\alias{rmse}
\title{Function to compute RMSE between true and filtered positions}
\usage{
rmse(true, filtered)
}
\arguments{
\item{true}{matrix of true positions (n x 2)}

\item{filtered}{matrix of filtered positions (n x 2)}
}
\value{
RMSE value (numeric)
}
\description{
Function to compute RMSE between true and filtered positions
}
