% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_models.R
\name{rmvt_mixture}
\alias{rmvt_mixture}
\title{Sample from mixture of multivariate student for error in ARGOS positions}
\usage{
rmvt_mixture(n, params)
}
\arguments{
\item{n}{number of samples}

\item{params}{parameters a, df, sigma_obs, rho and p in the distribution (see Brost et al. 2015).}
}
\value{
matrix of size n x 2 with samples
}
\description{
Sample from mixture of multivariate student for error in ARGOS positions
}
