% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_models.R
\name{rscaledt}
\alias{rscaledt}
\title{Generate random samples from scaled t-distribution}
\usage{
rscaledt(n, scale, df)
}
\arguments{
\item{n}{number of samples}

\item{scale}{positive scale parameter}

\item{df}{degree of freedom}
}
\description{
Generate random samples from scaled t-distribution
}
