% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_2D_trajectory}
\alias{simulate_2D_trajectory}
\title{Simulate penalised Langevin SDE trajectory in two dimensions.}
\usage{
simulate_2D_trajectory(
  n_sim,
  sde_params,
  potential_params = NULL,
  error_params,
  error_dist,
  polygon,
  lambda,
  U0,
  N,
  dt,
  scheme = "Lie-Trotter",
  split_around_fixed_point = TRUE,
  ncores = parallel::detectCores() - 1,
  seed = 42
)
}
\arguments{
\item{n_sim}{number of trajectories to simulate}

\item{sde_params}{List of movement parameters tau,nu,omega}

\item{potential_params}{List of parameters for the potential.
Only necessary when fixed_point_ind is not null, and only works with 
mixture of Gaussian potentials. The list has elements 
\code{potential_params$alpha} which is a vector of weights, 
\code{potential_params$B} which is a list of covariance matrices, 
and \code{potential_params$x_star} which is a matrix of coordinates of each 
center in the mixture.}

\item{error_params}{List of parameters of the measurement error distribution}

\item{error_dist}{Distribution of the measurement error. 
Options are normal,scaled_t and argos}

\item{polygon}{A matrix defining the boundary polygon (n x 2 coordinates)}

\item{lambda}{Penalization parameter.}

\item{U0}{initial state. Initial position and velocity for RACVM, 
and only initial position for BM and OU}

\item{N}{Integer, number of time steps to simulate}

\item{dt}{Numeric, time step size}

\item{scheme}{String, splitting scheme to use. Options are "Lie-Trotter" and "Strang"}

\item{split_around_fixed_point}{Boolean whether
to do the splitting adaptatively around a fixed point of the potential}

\item{ncores}{Number of cores to use for parallel simulation}

\item{seed}{Random seed for reproducibility}
}
\value{
A dataframe containing with Nxn_sim rows with columns
\itemize{
  \item X1, X2: true position
  \item V1, V2 : true velocity (for RACVM)
  \item Y1, Y2 : Noisy position
  \item ID : trajectory identifier
}
}
\description{
Simulate penalised Langevin SDE trajectory in two dimensions.
}
\details{
The function simulates movement according to a penalised Langevin SDE. When the 
animal hits the polygon boundary, a repulsive
force is applied. Observations include measurement error.
}
