% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gradient.R
\name{dRACVM_cov}
\alias{dRACVM_cov}
\title{Function to compute partial derivatives of RACVM covariance matrix Q (Delporte et .al 2025)}
\usage{
dRACVM_cov(tau, nu, omega, h)
}
\arguments{
\item{tau}{RACVM parameter tau}

\item{nu}{RACVM parameter nu}

\item{omega}{RACVM parameter omega}

\item{h}{time step}
}
\value{
List with three elements: derivatives of Q with respect to
#' tau, nu, and omega : three 4x4 matrices: dQ_dtau, dQ_dnu, dQ_domega
}
\description{
Function to compute partial derivatives of RACVM covariance matrix Q (Delporte et .al 2025)
}
