% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gradient.R
\name{llk_gradient}
\alias{llk_gradient}
\title{Compute gradient of log-likelihood over entire trajectory}
\usage{
llk_gradient(
  data,
  push_mat,
  potential_grad_mat,
  tau,
  nu,
  omega,
  error_dist = NULL,
  error_params = NULL,
  scheme = "Lie-Trotter",
  ind_fixed_point = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame with columns X1, X2, V1, V2, time
containing "true" positions and velocities at each time step}

\item{push_mat}{Matrix of shape (n_steps-1, 2) containing the push
at each time step}

\item{potential_grad_mat}{Matrix of shape (n_steps-1, 2) containing the
gradient of the potential at each time step}

\item{tau}{RACVM parameter tau}

\item{nu}{RACVM parameter nu}

\item{omega}{RACVM parameter omega}

\item{error_dist}{Error distribution (not used here)}

\item{error_params}{Error distribution parameters (not used here)}

\item{scheme}{Integration scheme (not used here)}

\item{ind_fixed_point}{Indices of fixed points (not used here)}

\item{verbose}{If TRUE, print detailed computation steps}
}
\value{
Numeric vector of length 3 with gradients w.r.t. tau, nu, omega
}
\description{
Compute gradient of log-likelihood over entire trajectory
}
