% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llk_gradient.R
\name{llk_gradient_one_step}
\alias{llk_gradient_one_step}
\title{Compute gradient of log-likelihood for one time step (works only for Lie-Trotter scheme 
with splitting around 0)}
\usage{
llk_gradient_one_step(
  U_next,
  U_prev,
  delta,
  push,
  potential_grad,
  tau,
  nu,
  omega,
  error_dist = NULL,
  error_params = NULL,
  scheme = "Lie-Trotter",
  ind_fixed_point = NULL,
  verbose = FALSE
)
}
\arguments{
\item{U_next}{State vector at next time step (X1, X2, V1, V2)}

\item{U_prev}{State vector at previous time step (X1, X2, V1, V2)}

\item{delta}{Time step size}

\item{push}{Push vector at previous time step (length 2)}

\item{potential_grad}{Gradient of potential at previous time step (length 2)}

\item{tau}{RACVM parameter tau}

\item{nu}{RACVM parameter nu}

\item{omega}{RACVM parameter omega}

\item{error_dist}{Error distribution (not used for the moment)}

\item{error_params}{Error distribution parameters (not used for the moment)}

\item{scheme}{Integration scheme (not used here)}

\item{ind_fixed_point}{Indices of fixed points (not used for the moment)}

\item{verbose}{If TRUE, print detailed computation steps}
}
\value{
Numeric vector of length 3 with gradients w.r.t. tau, nu, omega
}
\description{
Compute gradient of log-likelihood for one time step (works only for Lie-Trotter scheme 
with splitting around 0)
}
