% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{particle_filter2D}
\alias{particle_filter2D}
\title{Particle filter for movement in 2D with polygon constraint based on a penalization term}
\usage{
particle_filter2D(
  data,
  sde_params,
  potential_params = NULL,
  error_params,
  error_dist = "normal",
  polygon,
  U0,
  lambda,
  num_particles,
  scheme = "Lie-Trotter",
  split_around_fixed_point = FALSE,
  ESS_threshold = 0.5,
  verbose = FALSE
)
}
\arguments{
\item{data}{A data.frame containing observations with columns: time, Y1, Y2}

\item{sde_params}{List of parameters: RACVM parameters mu1, mu2,tau, nu, omega 
OU parameter : mu1,mu2,sigma,beta}

\item{potential_params}{Parameters of mixture of gaussian potential. Only necessary
when split_around_fixed_point=TRUE}

\item{error_params}{List of parameters of the measurement error distribution.
sigma_obs for normal, scale and df for scaled_t,
 sigma_obs,rho, a, df and p for argos.}

\item{error_dist}{distribution of the measurement noise. Options are
normal,scaled_t for student-t as in Weensveen, 
and argos for multivariate student as in Brost et .al.}

\item{polygon}{A matrix defining the boundary polygon (n x 2 coordinates)}

\item{U0}{initial state}

\item{lambda}{penalization parameter}

\item{num_particles}{Integer, number of particles to use in the filter}

\item{scheme}{Either Lie-Trotter or Strang for the numerical approximation of 
the underlying dynamic}

\item{split_around_fixed_point}{Whether to split the SDE adaptatively around fixed point
of the drift. Only works for mixture of gaussian potential.}

\item{verbose}{Logical, whether to print progress messages (default: FALSE)}
}
\value{
A list containing:
\itemize{
  \item particles: Array of particle states (num_particles x num_states x N)
  \item weights: Matrix of particle weights (num_particles x N)
  \item total_weights: Vector of sum unnormalized weights at each time step
  \item loglik: Estimated log-likelihood of the data
  \item ancestors: Matrix of ancestor indices for each particle at each time step
  \item push_array: Array of inward push values for each particle at each time step
  \item ind_fixed_point_mat: Matrix of fixed point indices for each particle at each time step
  
}
}
\description{
Particle filter for movement in 2D with polygon constraint based on a penalization term
}
