% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{plot_trajectory_in_polygon}
\alias{plot_trajectory_in_polygon}
\title{Visualizes both true and observed animal movement trajectories 
within a specified polygon boundary.}
\usage{
plot_trajectory_in_polygon(
  data,
  polygon,
  true_opacity = 0.2,
  obs_opacity = 0.2,
  potential_opacity = 0.4,
  true_size = 0.2,
  obs_size = 0.1,
  show_obs = TRUE,
  potential_grid = NULL,
  n_sub = 1,
  zoom_box = NULL
)
}
\arguments{
\item{data}{df from `simulate_2D_trajectory()`. If multiple IDs, individual trajectories are on individual facets}

\item{polygon}{SpatialPolygons object defining the boundary (from `sp` package)}

\item{true_opacity}{Opacity for true trajectory line (0-1, default: 0.5)}

\item{obs_opacity}{Opacity for observed trajectory line (0-1, default: 0.5)}

\item{potential_opacity}{Opacity for potential grid (0-1, default: 0.4)}

\item{true_size}{Size of points/lines for true trajectory (default: 0.2)}

\item{obs_size}{Size of points/lines for observed trajectory (default: 0.1)}

\item{show_obs}{Logical indicating whether to show observed trajectory (default: TRUE)}

\item{potential_grid}{spatial grid with columns x, y and H for the underlying 
potential driving the movement}

\item{n_sub}{Integer indicating subsampling rate for plotting 
(default: 1, no subsampling)}

\item{zoom_box}{should be a named list: list(xmin, xmax, ymin, ymax)}
}
\value{
A ggplot object showing:
\itemize{
  \item Polygon boundary (light blue fill)
  \item True trajectory (black line and points)
  \item Observed trajectory (red line and points with measurement error)
}
}
\description{
Visualizes both true and observed animal movement trajectories 
within a specified polygon boundary.
}
\details{
The plot helps visualize the relationship between true movement and noisy observations.
}
