% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SGD.R
\name{SGD_Fisher}
\alias{SGD_Fisher}
\title{Stochastic Gradient Descent with Fisher Information Preconditioning for SDE parameter estimation 
(Baey et .al 2023). Only works with Naive (split around 0) Lie-Trotter scheme for now.}
\usage{
SGD_Fisher(
  data,
  sde_params,
  fixpar = NULL,
  SGD_iter,
  potential_params = NULL,
  error_dist = NULL,
  error_params = NULL,
  scheme = "Lie-Trotter",
  polygon,
  U0,
  lambda,
  num_particles,
  split_around_fixed_point = FALSE,
  verbose = FALSE,
  gamma0 = 1e-04,
  K_preheat = 1000,
  alpha = 2/3,
  smoothing = FALSE
)
}
\arguments{
\item{data}{Data frame with observed data}

\item{sde_params}{Named list of initial SDE parameters}

\item{fixpar}{Character vector of parameter names to fix during optimization}

\item{SGD_iter}{Number of SGD iterations}

\item{potential_params}{Parameters ofr the mixture of gaussian potential. Only
needed when split_around_fixed_point=TRUE}

\item{error_dist}{Error distribution for the observation model}

\item{error_params}{Parameters for this error distribution}

\item{scheme}{Splitting scheme ("Lie-Trotter" or "Strang")}

\item{polygon}{Polygon defining the domain of interest}

\item{U0}{Initial position and velocity vector}

\item{lambda}{Penlization parameter}

\item{num_particles}{Number of particles for the particle filter}

\item{split_around_fixed_point}{Logical indicating whether to split the potential around its fixed point}

\item{verbose}{Logical indicating whether to print progress messages}

\item{gamma0}{Initial step size parameter}

\item{K_preheat}{Number of preheating iterations}

\item{alpha}{Step size decay parameter}

\item{smoothing}{Logical indicating whether to use smoothing when inputting a latent trajectory}
}
\value{
A list containing:
#' - theta: Matrix of parameter estimates at each iteration
#' - Delta: Matrix of running gradient estimates
}
\description{
Stochastic Gradient Descent with Fisher Information Preconditioning for SDE parameter estimation 
(Baey et .al 2023). Only works with Naive (split around 0) Lie-Trotter scheme for now.
}
